Add basic US services to the categories existing.
This could be broken down by state when it becomes unwieldy and utilities could
be further separated into water, gas, electric if desired.  Let me know if there
is anything specifically incorrect in how I'm putting this data in.

This is mostly a collection of services I've used in the past and I'll add to
them as I review old documents.